Move HTTP header methods APIUtils -> Access since that is the only place they are used
